+++ pyi.py
@@ -147,6 +147,21 @@ _BAD_Y027_IMPORTS = {
     "Container": "T",
 }
 
+_Y026_TYPEALIAS_BLACKLIST = frozenset(
+    {
+        "Literal",
+        "Union",
+        "Optional",
+        "Annotated",
+        "dict",
+        "tuple",
+        "list",
+        "set",
+        "frozenset",
+        "type",
+    }
+)
+
 
 class PyiAwareFlakesChecker(FlakesChecker):
     def deferHandleNode(self, node: ast.AST | None, parent) -> None:
@@ -326,7 +341,6 @@ _is_TypeAlias = partial(_is_object, name="TypeAlias", from_=_TYPING_MODULES)
 _is_NamedTuple = partial(_is_object, name="NamedTuple", from_={"typing"})
 _is_TypedDict = partial(_is_object, name="TypedDict", from_=_TYPING_MODULES)
 _is_Literal = partial(_is_object, name="Literal", from_=_TYPING_MODULES)
-_is_Union = partial(_is_object, name="Union", from_={"typing"})
 _is_abstractmethod = partial(_is_object, name="abstractmethod", from_={"abc"})
 _is_Any = partial(_is_object, name="Any", from_={"typing"})
 _is_overload = partial(_is_object, name="overload", from_={"typing"})
@@ -814,6 +828,7 @@ class PyiVisitor(ast.NodeVisitor):
             self.generic_visit(node)
         if target_name is None:
             return
+        assert isinstance(target, ast.Name)
         assignment = node.value
         if isinstance(assignment, ast.Call):
             self._check_assignment_to_function(
@@ -831,7 +846,7 @@ class PyiVisitor(ast.NodeVisitor):
             return self._Y015_error(node)
 
         if not is_special_assignment:
-            self._check_for_type_aliases(node, target_name, assignment)
+            self._check_for_type_aliases(node, target, assignment)
 
     def visit_AugAssign(self, node: ast.AugAssign) -> None:
         """Allow `__all__ += ['foo', 'bar']` in a stub file"""
@@ -843,8 +858,19 @@ class PyiVisitor(ast.NodeVisitor):
         else:
             self.visit(value)
 
+    def _Y026_error(
+        self, node: ast.Assign, target: ast.Name, assignment: ast.expr
+    ) -> None:
+        new_node = ast.AnnAssign(
+            target=target,
+            annotation=ast.Name(id="TypeAlias", ctx=ast.Load()),
+            value=assignment,
+            simple=1,
+        )
+        self.error(node, Y026.format(suggestion=unparse(new_node)))
+
     def _check_for_type_aliases(
-        self, node: ast.Assign, target_name: str, assignment: ast.expr
+        self, node: ast.Assign, target: ast.Name, assignment: ast.expr
     ) -> None:
         """
         Check for assignments that look like they could be type aliases,
@@ -859,13 +885,17 @@ class PyiVisitor(ast.NodeVisitor):
         whether these are type aliases or variable aliases,
         unless you're a type checker (and we're not).
         """
-        if isinstance(assignment, ast.BinOp):
-            return self.error(node, Y026)
+        if isinstance(assignment, ast.BinOp) or _is_Any(assignment):
+            return self._Y026_error(node, target, assignment)
         if not isinstance(assignment, ast.Subscript):
             return
-        subscripted_object = assignment.value
-        if _is_Union(subscripted_object) or _is_Literal(subscripted_object):
-            self.error(node, Y026)
+
+        name_if_from_builtins_or_typing = _get_name_of_class_if_from_modules(
+            assignment.value, modules=_TYPING_MODULES | {"builtins"}
+        )
+
+        if name_if_from_builtins_or_typing in _Y026_TYPEALIAS_BLACKLIST:
+            self._Y026_error(node, target, assignment)
 
     def visit_Name(self, node: ast.Name) -> None:
         self.all_name_occurrences[node.id] += 1
@@ -1652,7 +1682,7 @@ Y025 = (
     'Y025 Use "from collections.abc import Set as AbstractSet" '
     'to avoid confusion with "builtins.set"'
 )
-Y026 = "Y026 Use typing_extensions.TypeAlias for type aliases"
+Y026 = 'Y026 Use typing_extensions.TypeAlias for type aliases, e.g. "{suggestion}"'
 Y027 = "Y027 Use {good_syntax} instead of {bad_syntax} (PEP 585 syntax)"
 Y028 = "Y028 Use class-based syntax for NamedTuples"
 Y029 = "Y029 Defining __repr__ or __str__ in a stub is almost always redundant"