+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 Features:
 * Introduce Y041: Ban redundant numeric unions (`int | float`, `int | complex`,
   `float | complex`).
+* Improve error message for Y026 check.
+* Expand Y026 check. Since version 22.4.0, this has only emitted an error for
+  assignments to `typing.Literal`, `typing.Union`, and PEP 604 unions. It now also
+  emits an error for assignments to `Annotated`, `Optional` and `Any`, as well as
+  subscripted `tuple`s, `dict`s, `set`s, `frozenset`s, `list`s, and `type`s.
 
 ## 22.5.1